Adicionando endpoint para buscar estatisticas da transação e documentando a api com swagger openapi.

+++ src/main/java/io/github/leolimaferreira/desafio_itau_junior/model/Transacao.java
@@ -1,14 +1,12 @@
 package io.github.leolimaferreira.desafio_itau_junior.model;
 
-import java.math.BigDecimal;
-import java.time.LocalDateTime;
-import java.time.OffsetDateTime;
-
 import lombok.Data;
+
+import java.time.OffsetDateTime;
 
 @Data
 public class Transacao {
 	
-	private BigDecimal valor;
+	private Double valor;
 	private OffsetDateTime dataHora;
 }
